Boilerplate for html and css
Starting html and css for the project. The CSS is at the moment minify because I am writing the css with Sass. If is a problem I can change it back to normal css. In the css I have added basic styles for the css as a normalizer (So it looks similar over all modern web browsers) a minor reset and font base to 10px (to use rem units instead of pixels)